refactor: pageのcss追加置き換え
