+++ src/app/Top.css.ts
@@ -1,0 +1,8 @@
+import { style } from '@vanilla-extract/css';
+
+export const styles = {
+  baseHeading1Articles: style({
+    padding: '0 0 10px',
+    borderBottom: '2.5px solid lightgray'
+  })
+};
