--- src/app/Top.module.css
@@ -1,4 +1,0 @@
-.baseHeading1Articles {
-  padding: 0 0 10px;
-  border-bottom: 2.5px solid lightgray;
-}
