+++ src/app/profile/Profile.css.ts
@@ -1,0 +1,17 @@
+import { style } from '@vanilla-extract/css';
+
+export const styles = {
+  contents: style({
+    marginTop: '1.25em',
+    marginRight: 'auto',
+    marginLeft: 'auto',
+    marginBottom: '2em'
+  }),
+  baseHeading1Profile: style({
+    padding: '0 0 10px',
+    borderBottom: '2.5px solid lightgray'
+  }),
+  baseHeading2Profile: style({
+    fontSize: '1.25em'
+  })
+};
